fix: patches now check for "E Commerce Settings"
ordering of patch execution also changed as one of the patches (` make_homepage_products_website_items` ) depended on another patch being executed before it ( `add_homepage_field` )
