+++ webshop/setup/install.py
@@ -4,6 +4,7 @@ import frappe
 from frappe import _
 from frappe.custom.doctype.custom_field.custom_field import create_custom_fields
 
+from webshop.webshop.utils.setup import has_ecommerce_fields
 
 def after_install():
 	run_patches()
@@ -46,19 +47,6 @@ def copy_from_ecommerce_settings():
 		)
 
 		query.run()
-
-def has_ecommerce_fields() -> bool:
-	table = frappe.qb.Table("tabSingles")
-	query = (
-		frappe.qb.from_(table)
-		.select(table.field)
-		.where(table.doctype == "E Commerce Settings")
-		.limit(1)
-	)
-
-	data = query.run(as_dict=True)
-	return bool(data)
-
 
 def drop_ecommerce_settings():
 	frappe.delete_doc_if_exists("DocType", "E Commerce Settings", force=True)
@@ -229,12 +217,12 @@ def say_thanks():
 patches = [
 	"create_website_items",
 	"populate_e_commerce_settings",
+	"add_homepage_field",
 	"make_homepage_products_website_items",
 	"fetch_thumbnail_in_website_items",
 	"convert_to_website_item_in_item_card_group_template",
 	"shopping_cart_to_ecommerce",
 	"copy_custom_field_filters_to_website_item",
-	"add_homepage_field"
 ]
 
 def run_patches():
